fix parameter type in f_pc_pause

+++ include/f/f_pc/f_pc_pause.h
@@ -9,7 +9,7 @@ extern "C" {
 
 s32 fpcPause_IsEnable(void *pProc, u8 expected);
 s32 fpcPause_Enable(void *pProc, u8 pauseMask);
-s32 fpcPause_Disable(void *pProc, u32 pauseMask);
+s32 fpcPause_Disable(void *pProc, u8 pauseMask);
 void fpcPause_Init(void *pProc);
 
 }
